added few objects, modified COntroller
objects affect differently players speed

+++ Assets/Scripts/Objects/ModifySpeed.cs
@@ -6,7 +6,7 @@ public class ModifySpeed : MonoBehaviour
 {
   private Rigidbody2D body;
   public Transform ObjectPosition;
-  public int speed_modifier;
+  public float speed_modifier;
 
   void Start()
   {
